refactor ResetUser and UserLink dialogs
