+++ ui/src/dialogs/users/DeleteUser.js
@@ -1,7 +1,6 @@
-import React, { useState } from 'react'
-import { Link, redirect, useLoaderData, useNavigate, useRouteLoaderData } from 'react-router-dom'
+import React from 'react'
+import { useNavigate } from 'react-router-dom'
 
-import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import Dialog from '@mui/material/Dialog'
 import DialogActions from '@mui/material/DialogActions'
