+++ ui/src/pages/Users.js
@@ -3,13 +3,7 @@ import { Link, redirect, useLoaderData, useNavigate, useRouteLoaderData } from '
 
 import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
-import ContentCopyRoundedIcon from '@mui/icons-material/ContentCopyRounded'
 import DeleteRoundedIcon from '@mui/icons-material/DeleteRounded'
-import Dialog from '@mui/material/Dialog'
-import DialogActions from '@mui/material/DialogActions'
-import DialogTitle from '@mui/material/DialogTitle'
-import DialogContent from '@mui/material/DialogContent'
-import DialogContentText from '@mui/material/DialogContentText'
 import Divider from '@mui/material/Divider'
 import EditRoundedIcon from '@mui/icons-material/EditRounded'
 import HomeRoundedIcon from '@mui/icons-material/HomeRounded'
@@ -20,7 +14,6 @@ import LockResetRoundedIcon from '@mui/icons-material/LockResetRounded'
 import Paper from '@mui/material/Paper'
 import PersonAddAltRoundedIcon from '@mui/icons-material/PersonAddAltRounded'
 import RefreshRoundedIcon from '@mui/icons-material/RefreshRounded'
-import TextField from '@mui/material/TextField'
 import Table from '@mui/material/Table'
 import TableBody from '@mui/material/TableBody'
 import TableCell from '@mui/material/TableCell'
@@ -29,11 +22,14 @@ import TableHead from '@mui/material/TableHead'
 import TableRow from '@mui/material/TableRow'
 import Typography from '@mui/material/Typography'
 
-import { getApps, getUsers, resetUser } from '../apis/uidataAPI'
+import { getApps, getUsers } from '../apis/uidataAPI'
 import Roles, { InsufficientRoles } from '../components/Roles'
 import NewUser from '../dialogs/users/NewUser'
 import DeleteUser from '../dialogs/users/DeleteUser'
 import EditUser from '../dialogs/users/EditUser'
+import UserLink from '../dialogs/users/UserLink'
+import ResetUser from '../dialogs/users/ResetUser'
+
 export async function userListLoader() {
   try {
     const userList = await getUsers()
@@ -116,21 +112,12 @@ export default function Users ({ setSuccessMessage, setErrorMessage }) {
     setDeleteUserDialogOpen(false)
   }
 
-  async function userLinkCopy () {
-    const activationInfo = `Username:\n${userLinkUser ? userLinkUser.username : ''}\nActivation Link:\n${userLinkUser ? `${window.location.href.replace('/users', '/account/')}${userLinkUser.linkID}` : ''}`
-    await navigator.clipboard.writeText(activationInfo)
-
-    // TODO - this success message causes a redraw, which causes the loader to fire again and load the new user.
-    // When we then close the dialog, we load again
-    // setSuccessMessage('Info copied')
-  }
-
   function openUserLink (user) {
     setUserLinkUser(user)
     setUserLinkDialogOpen(true)
   }
 
-  function userLinkDialogClose () {
+  function closeUserLink () {
     setUserLinkDialogOpen(false)
     if (userLinkTriggerRefresh) {
       setUserLinkTriggerRefresh(false)
@@ -138,82 +125,17 @@ export default function Users ({ setSuccessMessage, setErrorMessage }) {
     }
   }
 
-  function resetUserDialogOpen (user) {
+  function openResetUser (user) {
     setResetUser(user)
     setResetUserDialogOpen(true)
   }
 
-  function resetUserDialogClose () {
+  function closeResetUser () {
     setResetUserDialogOpen(false)
   }
 
-  async function resetUserAction () {
-    setLoading(true)
-    setResetUserDialogOpen(false)
-
-    try {
-      const user = await resetUser(resetUserUser.id)
-      openUserLink(user)
-      setUserLinkTriggerRefresh(true)
-    } catch (error) {
-      setErrorMessage(error.message)
-    }
-
-    setLoading(false)
-  }
-
-  const UserLinkDialog = (<Dialog open={userLinkOpen} onClose={userLinkDialogClose} aria-labelledby='user link'>
-    <DialogTitle id='user-link-dialog-title' className='dialog-top'>User Activation Link</DialogTitle>
-    <DialogContent>
-      <Box sx={{padding: '10px 0px'}}>
-        <Typography variant='body1' >Copy the information below and send it to the new user.  They can then<br/>activate their account and set their own password</Typography>
-      </Box>
-      <Box className='data-box' >
-        <Box sx={{padding: '10px'}}>
-          <Typography variant='body1' >Username:</Typography>
-          <Typography variant='body2' >{userLinkUser ? userLinkUser.username : ''}</Typography>
-        </Box>
-        <Box sx={{padding: '10px'}}>
-          <Typography variant='body1' >Activation Link:</Typography>
-          <Typography variant='body2' >{userLinkUser ? `${window.location.href.replace('/users', '/account/')}${userLinkUser.linkID}` : ''}</Typography>
-        </Box>
-      </Box>
-      <Box sx={{ display: 'flex', justifyContent: 'left', alignItems: 'center', padding: '10px 0px 0px 0px'}}>
-        <Button aria-label='Copy' variant='outlined' startIcon={<ContentCopyRoundedIcon />} onClick={userLinkCopy}>Copy Info</Button>
-      </Box>
-    </DialogContent>
-    <DialogActions>
-      <Button onClick={userLinkDialogClose} variant='contained' color='primary'>Close</Button>
-    </DialogActions>
-  </Dialog>
-  )
-
-  const ResetUserDialog = (
-    <Dialog open={resetUserOpen} onClose={resetUserDialogClose} aria-labelledby='reset user'>
-      <DialogTitle id='reset-user-dialog-title' className='dialog-top'>Reset User</DialogTitle>
-      <DialogContent>
-        <br/>
-        <DialogContentText>Username</DialogContentText>
-        <Box sx={{padding: '10px 0px'}}>
-          <TextField id='reset-username' disabled={true} name='reset-username' sx={{ width: '35ch' }} value={resetUserUser ? resetUserUser.username : ''} variant='outlined' />
-        </Box>
-        <Box sx={{padding: '10px 0px'}}>
-          <Typography variant='body1' >This will reset the user's password and generate a new activation link.<br />
-                                       Once you have sent them their new link, they will be able to set a new<br />
-                                       password for their account<br /><br />
-                                       Are you sure you want to reset the account?</Typography>
-        </Box>
-      </DialogContent>
-      <DialogActions>
-        <Button onClick={resetUserDialogClose} variant='outlined' color='primary'>Cancel</Button>
-        <Button onClick={resetUserAction} variant='contained' color='primary'>Reset</Button>
-      </DialogActions>
-    </Dialog>
-  )
-
-  let ui = ( <InsufficientRoles /> )
   if (Roles.roleCheck(userInfo.roles, [Roles.ADMIN])) {
-    ui = (
+    return (
       <Box padding='5px' sx={{ display: 'flex', flexDirection: 'column', height: '100%' }}>
         <Box padding='10px 10px 5px 10px'>
           <Box sx={{ display: 'flex' }}>
@@ -299,7 +221,7 @@ export default function Users ({ setSuccessMessage, setErrorMessage }) {
                             ?
                             null
                             :
-                            <IconButton size='small' edge='start' color='inherit' aria-label='menu' sx={{ mr: 2 }} onClick={() => { resetUserDialogOpen(user) }} >
+                            <IconButton size='small' edge='start' color='inherit' aria-label='menu' sx={{ mr: 2 }} onClick={() => { openResetUser(user) }} >
                               <LockResetRoundedIcon />
                             </IconButton>
                           }
@@ -321,11 +243,12 @@ export default function Users ({ setSuccessMessage, setErrorMessage }) {
         { newUserOpen ? <NewUser setLoading={setLoading} apps={apps} usernames={usernames} openUserLink={openUserLink} setUserLinkTriggerRefresh={setUserLinkTriggerRefresh} closeDialog={closeNewUser} setSuccessMessage={setSuccessMessage} setErrorMessage={setErrorMessage} /> : null }
         { deleteUserOpen ? <DeleteUser setLoading={setLoading} user={deleteUserUser} closeDialog={closeDeleteUser} setSuccessMessage={setSuccessMessage} setErrorMessage={setErrorMessage} />: null }
         { editUserOpen ? <EditUser setLoading={setLoading} user={editUserUser} apps={apps} closeDialog={closeEditUser} setSuccessMessage={setSuccessMessage} setErrorMessage={setErrorMessage} /> : null }
-        { userLinkOpen ? UserLinkDialog : null }
-        { resetUserOpen ? ResetUserDialog : null }
+        { userLinkOpen ? <UserLink user={userLinkUser} closeDialog={closeUserLink} setSuccessMessage={setSuccessMessage} setErrorMessage={setErrorMessage} /> : null }
+        { resetUserOpen ? <ResetUser setLoading={setLoading} user={resetUserUser} closeDialog={closeResetUser} openUserLink={openUserLink} setUserLinkTriggerRefresh={setUserLinkTriggerRefresh} setSuccessMessage={setSuccessMessage} setErrorMessage={setErrorMessage} /> : null }
       </Box>
     )
+  } else {
+    return ( <InsufficientRoles /> )
   }
-  return ui
 }
